test: split configuration.test.js in separated test files
